Remove duplicated code & bug fix

+++ Entities/Expenses.cs
@@ -11,7 +11,7 @@ namespace Entities
         public int GoodsId { get; set; }
         public int UnitId { get; set; }
         public decimal Price { get; set; }
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public DateTime Date { get; set; }
         public string ToCsv(int id)
         {
